Update mir-verifier and libraries to Rust nightly-2020-03-22.

On the mir-verifier side, the one major change is better constant handling.  Newer versions of rustc have gotten much more aggressive in applying constant propagation, meaning more kinds of values can show up in `Constant` operands.  There are also some minor changes to the overrides in `Mir.TransCustom`, mostly to account for renaming of items in the standard library.

The libraries have been updated, and the patches have been redone mostly from scratch to take advantage of recent improvements in our handling of raw pointers.  Only a few of the larger pieces, such as the patches to `slice::sort` or the `core::fmt` module, were copied over unmodified from the previous version.

The corresponding mir-json changes are in GaloisInc/mir-json#16.

+++ crux-mir/lib/libc/ci/dox.sh
@@ -13,11 +13,19 @@ PLATFORM_SUPPORT=platform-support.md
 rm -rf $TARGET_DOC_DIR
 mkdir -p $TARGET_DOC_DIR
 
+if ! rustc --version | grep -E "nightly" ; then
+    echo "Building the documentation requires a nightly Rust toolchain"
+    exit 1
+fi
+
+rustup component add rust-src
+cargo +nightly install cargo-xbuild
+
 # List all targets that do currently build successfully:
 # shellcheck disable=SC1003
 grep '[\d|\w|-]* \\' ci/build.sh > targets
 sed -i.bak 's/ \\//g' targets
-grep '^[_a-zA-Z0-9-]*$' targets > tmp && mv tmp targets
+grep '^[_a-zA-Z0-9-]*$' targets | sort > tmp && mv tmp targets
 
 # Create a markdown list of supported platforms in $PLATFORM_SUPPORT
 rm $PLATFORM_SUPPORT || true
@@ -39,10 +47,13 @@ while read -r target; do
 
     rustup target add "${target}" || true
 
+    # Enable extra configuration flags:
+    export RUSTDOCFLAGS="--cfg freebsd11"
+
     # If cargo doc fails, then try xargo:
     if ! cargo doc --target "${target}" \
              --no-default-features  --features extra_traits ; then
-        xargo doc --target "${target}" \
+        cargo xdoc --target "${target}" \
               --no-default-features  --features extra_traits
     fi
 
@@ -61,10 +72,3 @@ set -x
 
 # Copy the licenses
 cp LICENSE-* $TARGET_DOC_DIR/
-
-# If we're on travis, not a PR, and on the right branch, publish!
-if [ "$TRAVIS_PULL_REQUEST" = "false" ] && [ "$TRAVIS_BRANCH" = "master" ]; then
-  pip install ghp_import --install-option="--prefix=$HOME/.local"
-  "${HOME}/.local/bin/ghp-import" $TARGET_DOC_DIR
-  git push -qf "https://${GH_TOKEN}@github.com/${TRAVIS_REPO_SLUG}.git" gh-pages
-fi
